refactor(gateway): make contracts modules private (_-prefixed)
Rename contracts/{envelope,pagination,security}.py to _envelope/_pagination/
_security.py and re-export the public surface via __init__.py, matching the
repo's SPI convention (spi/auth/_models.py, spi/authn/_models.py). Callers
already import from the package path, so future internal refactors stay
contained. No behavior change.

Addresses review feedback on PR #345.
