+++ src/gateway/src/gateway/community/adapters/web/contracts/_envelope.py
