+++ src/gateway/src/gateway/community/adapters/web/contracts/_pagination.py
