+++ src/gateway/src/gateway/community/adapters/web/contracts/_security.py
